Relocated cleanup back to mods.ts
